Update package structure
"ui" package is used to handle everything that has to do with the
user interface. Inside it, there will be further division in more
packages to bundle up the code in logical parts. Initially, the login
part of the application will reside inside "auth" and later there will
be new packages holding any other parts of the application.