Add auth fragments
Add all fragments related to the auth part of the application. The main
launcher fragment is the first thing the user interacts with. The other
fragments are responsible for what the name implies.

Add colours and some string resources to support all the new  fragments.

Add the initial classes that match those fragment without any
implementation yet.