Setup Dagger and Room
Add all the default classes that are always needed for Dagger.
Build the AppModule and the AuthModule along with everything that is
under them.
Initialize some example models and Daos for Room, subject to change
when the communication protocol is setup.
Create BaseActivity class to be extended by any Activity in order to
properly extend DaggerAppCompatActivity as dagger requires.
Add BaseAuthFragment to be extended by all fragments inside auth module,
which extends DaggerFragment and Injects the viewModel that all those
fragments share.
Use Inject on SessionManager, AuthRepository and AuthViewModel
constructors to inject relevant variables to them.