Create MVI architecture structure
Create the wrapper classes and generic implementations for Events etc
to be implemented in future classes.
AbsentLiveData is a simple class taken from Google's Samples that is
there to help with having a LiveData object that contains null value.
The BaseViewModel is there to enforce the MVI architecture by providing
a couple of abstract methods, along with some implemented ones that
will be identical for any ViewModel following the MVI pattern.
DataState class is a generic way to hold the data state. It is built to
contain all the relevant information on the state of the data, like if
there is an error, if it is in the loading state, and whether it
contains data or not.
The Data that it contains, is a simple data class that contains two
objects, both of which are wrapped around the Event wrapper class.
The Event wrapper class is used in order to provide us with some helper
methods like getting the content out of it, having two different
functions inside the companion object that assist us in creating it,
and most importantly providing us with a boolean that knows if we have
accessed the content of it or not. This is useful as later, when we are
observing it, we do not want to show the relevant toast message or
whatever it is more than once, as it happens when there is a
configuration change on the fragments and the observers are initialized
again and the Event is looked at again. Using this way, we ensure that
whatever triggers from the Event will only happen the first time.
The ResponseType class that is used as mentioned before is a simple way
to label the response in a way that then when observed on, we will know
in which way we want to show to the user this response, like a Toast
message, a Dialog or even None if there is no need for it.