SessionManager implementation
Update the session manager to include methods like login and logout.
They are responsible for updating the LiveData object containing the
cached authentication credentials. Logout is also responsible for
clearing the authentication credentials from the database.
Start initializing the Main part of the application as well. The
activity with a simple bottom navigation is initialized. The Dagger
scope is not yet initialized, however the MainActivity is contributed
inside the ActivityBuildersModule for it to work temporarily.
The two state events inside auth are created, namely Login and Register
attempt events.
Along with the state events, the view state class is also created,
being responsible for view state related information, like the fields
containing proper user input for registration and generally creating
LoginFields and RegisterFields classes, responsible for saving the
information on them.
Start subscribing with the activities on the relevant LiveData
regarding the authentication state and navigating from/to the auth
activity depending on the state of the authentication.
Give a testing behavior to the login button by simply setting the state
of the sessionManager to a fake authenticated state to test logging in
and going to the main a activity.