Implement real login/register and error message handling
Update model classes to follow the rules of open-api.xyz for now.
Subscribe the observer on auth activity to react to the data event
changing, and updating the auth token in case it does update, which in
turn triggers the other observer which looks at the viewState, which
calls the login method inside the sessionManager.
Add register and login functions inside the respective fragments, which
both simply update the state of the viewmodel, which is observed by the
auth activity and is then handled.
Create the attemptLogin and attemptRegister methods inside the
repository which call the openApiAuthService interface, and then handle
the response by temporarily for now turning it into a LiveData object
manually by using switchMap.
Create NetworkBoundResource class.
Implement NBR in AuthRepository.
Create extensions for showing toasts, dialogs etc.
Implement DataStateChangeListener on baseActivity in order to observe
it and act accordingly. Have to trigger it from inside the activities
and call it inside the observer on change.