Implement basic subscribe/publish mqtt functionality
Update Dagger to work for the Main part of the app as well.
Publish uses MVI pattern almost perfectly, some refinement may be
needed.
Subscribe currently just interacts with the ViewModel directly and
updates the viewState directly as well, as a proof of concept before
meeting #4 with the teachers.
Main Activity also implements a temporary fragment which is simply
inflated and is used to subscribe/publish to a true/false topic (proof
of concept for an on/off lamp for example).