Initialize device recycler view
Add Device sealed class to be extended by all types of devices so that
the recycler adapter can have a list of this super class and contain
all types of devices.
Add DeviceListAdapter, the class responsible for holding the list of
Devices and using the recyclerView to show them to the user. It
contains ViewHolders for all types of views so far, and is open for
extension as more device types are added to the Device model class.
Add all initial xml layouts for the current device types we have, along
with an empty one for the rest of them that are not finalized yet.