Populate recycler view from topics list
Add Topic class to contain all the hard-coded topics to the house.
Also contains a hash map from a topic String to the class that will
represent its state, for example: "xLight" -> Device.InteractiveOnOff.

Create builder method for Device, to manage constructing a Device
object from the appropriate class depending on the topic by using the
hash map from the Topic class.

Update repository to use the Device.builder to return the completed
dataState, which will then update the viewModel's viewState.

Recycler view now is updated each time there is a change to the
DeviceFields in the MainViewState, which is updated from the
UpdateDeviceListEvent event. The method setDevicesFields does not just
replace the old list, but it adds on top of the list all the new
devices that may have come on this change.