Allow recycler views to publish their results to the broker
Handle interaction from the viewHolders to the fragment through one
single "Interaction" interface, that simply sends the new state of the
device of that position. That is then taken and turned into an MQTT
message by calling the function asMqttMessage of the Device class. That
message is then published through the attemptPublish viewState state.

Fix bug with views jumping to the bottom of the list due to bad way of
updating the list inside the addDevice method of DeviceFields.