Add subscription to MQTT broker and update recycler view with it
Subscription of the topics happens at the MainViewModel and for now
does break the MVI pattern. This is done due to the fact that observing
the topic is a constant state, while so far we have used MVI to do
one-shot calls that immediately respond with data or a fail response
and are done forever. I have not found a way to embed this observable
into the pattern so far, something to be taken into consideration later
down the line.

For now the subscription happens on main fragment initialization, and
does not check in any way later it the connection has been dropped or
not. In a future iteration the connection problems can be addressed. As
a temporary solution, it could be at least moved to a later stage of
the fragment's lifecycle so that it can re-subscribe along with
re-connecting to the MQTT broker when the screen is resumed (like after
the phone was locked for a bit).

As discussed here https://stackoverflow.com/a/50062174/9440211 we are
doing a "hacky" way of forcing the recycler view to update. We are
overriding submitList to fix problem with the new list being == to the
old list, which is true because we are also updating the local variable
whenever there is a change. We get around it by sending a copy of the
list which does not return true on the == operation and therefore
continues on the method and triggers the notifyDataSetChanged.