Update MQTT topics to the new protocol proposed at 24/11/2019
Remove the hard-coded initializations, now gets which views will be
inflated by subscribing to smarthouse/# and getting all retained
messages.

Change callback to instead of directly manipulating the ViewState
directly, to creating a new StateEvent (Intent). I am not sure if that
better follows the MVI pattern, since it is not a user intent, but
there is always room for fixing it later once a better way is found.

Note that with the new protocol and the new devices coming, the XML
files are not created yet and some temporary ones were placed as
placeholders. This will be fixed in future commits.