Update topic to only use the first two parameters for the recycler view
Due to the fact that the protocol now makes it so that one device may
have more than one topic, with the last parameter representing
different things for the same device, the view was then inflating two
separate instances of the same device for each unique topic.

This commit updates the functionality so that the views are inflated
only by depending on the first two parameters of the topic, example
smarthouse/lamp/state and smarthouse/lamp/brightness will inflate one
view only, being smarthouse/lamp. The differ list on the listAdapter is
given only one device to inflate.

The merging of the devices happens on a previous level. Now when the
device is built when calling deviceBuilder() inside Device.kt it
returns the device with the topic containing only the first two
parameters. Then when the device lists are merged called on
setDevicesFields inside the MainViewModel, the functionality of
addDevice has been updated to merge the old version and the new version
of the device together.

Now when getting user interaction on the recycler view, the main topic
shall be used as a base, and the final parameter will need to be added
depending what the user is interacting with.