Add device populating button
Since the online broker constantly loses the latest retained values,
this new button is responsible for simply adding all the devices to it
to make it easier to work with the app when not all of the devices are
retained.

Edit the primary color so that the dialog options are not invisible.