Modify filter dialog to manage more screen sizes.
Make better use of the Material Dialogs library by using the built-in
button and title creators to allow for the use of a scroll view for the
radio boxes. This makes it possible to properly display the information
to the user to almost all screen sizes except some occasions where the
height is so small that it would be impossible to do so. This fixes the
problem with the way it was showing before in landscape mode as well.